video_processing_RAG_System/requirements.txt
llama-index-vector-stores-lancedb
llama-index-multi-modal-llms-openai
llama-index-embeddings-clip
llama-index-readers-file
llama_index
openai-whisper
lancedb
moviepy
pytube
pydub
SpeechRecognition
ffmpeg-python
soundfile
torch
torchvision
matplotlib
scikit-image
ftfy
regex
tqdm
